requirements.txt
aniso8601==8.0.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
MyApplication==0.1.0
pycparser==2.20
PyJWT==1.4.2
PyMySQL==0.10.1
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
